requirements.txt
python-dotenv
google-cloud-speech
pydub